% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{load_individuals}
\alias{load_individuals}
\title{Construct a population}
\usage{
load_individuals(pid, pid_mom, pid_dad, progress = TRUE,
  error_on_pid_not_found = TRUE)
}
\description{
Construct a population
}
